% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{get_dfdesc}
\alias{get_dfdesc}
\title{Build a dataframe of basic descriptives.}
\usage{
get_dfdesc(.df)
}
\arguments{
\item{.df}{A data.frame}
}
\description{
\code{get_dfdesc}
}
\details{
This function is designed for use in the launch_varview GUI.
}
