% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{get_dflabs}
\alias{get_dflabs}
\title{Build a dataframe of variable names, classes, and labels.}
\usage{
get_dflabs(.df)
}
\arguments{
\item{.df}{A data.frame}
}
\description{
\code{get_dflabs}
}
\details{
This function is designed for use in the launch_varview GUI.
}
