% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{init_project}
\alias{init_project}
\title{Initiate a new Rproject from the console.}
\usage{
init_project(proj_name, create_folder = TRUE)
}
\arguments{
\item{proj_name}{a character string to be used as the name of the .Rproj file and directory}

\item{create_folder}{a logical value, whether to create a folder to hold the .Rproj file (TRUE) or
(FALSE) place the .Rproj folder directly in the current working directory.}
}
\description{
\code{init_project}
}
\details{
This function is designed for use in the launch_project GUI. By default,
it creates a folder inside the working directory with a name equal to the character
value given to the proj_name argument (this behavior can be suppressed by setting 
the argument create_folder = FALSE). A .Rproj file is then created and initialized 
inside the folder (or working directory, if create_folder = FALSE).
}
