% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_pmap.R
\name{launch_pmap}
\alias{launch_pmap}
\title{An interactive exploration of corresondence analysis objects.}
\usage{
launch_pmap(.ca, row_labels = NULL, col_labels = NULL,
  launch_type = "web")
}
\arguments{
\item{.ca}{A ca object, as returned from the package "ca".}

\item{row_labels}{A vector of row labels.}

\item{col_labels}{A vector of column labels.}
}
\description{
\code{launch_pmap} addional thing description
}
\details{
Insert description of uses
}
