% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_varview.R
\name{launch_varview}
\alias{launch_varview}
\title{An interactive dataframe explorer!}
\usage{
launch_varview(.df, launch_type = "web")
}
\arguments{
\item{.df}{A data.frame}
}
\description{
\code{launch_varview} addional thing description
}
\details{
The purpose of this gadget is to facilitate quick exploration of the variables contained
within a dataframe.
}
