% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbit.R
\name{orbit}
\alias{orbit}
\title{Calculate the rotated location of a point.}
\usage{
orbit(coords, rotation)
}
\arguments{
\item{coords}{A data frame containing the x,y coordinates of points to be rotated.}

\item{rotation}{A numeric value indicating the degreee of rotation.}
}
\value{
a data frame of the same length as coords, containing the rotated x,y coordinates for each point.
}
\description{
\code{orbit}
}
\details{
This function is designed for use in the launch_pmap GUI.
}
