% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_launch.R
\name{set_launch}
\alias{set_launch}
\title{Launch a GUI}
\usage{
set_launch(ui, server, launch_type)
}
\arguments{
\item{launch_type}{A data frame containing the x,y coordinates of points to be rotated.}
}
\description{
\code{set_launch}
}
\details{
This function is designed for use in the brownie GUIs.
}
